requirements.txt
Flask
python-docx
Flask-WTF